requirements.txt
flask
requests
chromadb
sentence_transformers
pandas
python-dotenv
pyyaml
oracledb
wcwidth
google-generativeai
